pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nit"
version = "0.1.0"
description = "Нить — AI-агент для поиска своего человека"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "aiogram>=3.13.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    "redis>=5.2.0",
    "arq>=0.26.0",
    "openai>=1.57.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.17",
    "aioboto3>=13.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
moderation = [
    "nudenet>=3.4.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.setuptools.packages.find]
where = ["."]
